Migrate python package to new bindings
The old ctypes-based interface was removed. The structure of the
repository was changed such that it can be easily installed using pip.
The python tests are passing, although I had to comment out several
parts. Pretty formatting is not supported beyond what the C++ version is
providing. Also, since the Sampler interface was refactored, several
tests were already broken before.