Use pyproject.toml, change directory structure
Metadata from setup.py was moved to pyproject.toml, more metadata was
added.

The alpaca python package is now in a subdirectory in the python
directory, as it was before, because this makes more sense.

The test directories were renamed to tests, because this name is more
commonly used (and more specific).